move operator to class
