+++ lang/src/org/partiql/lang/eval/physical/operators/UnpivotRelationalOperatorFactory.kt
@@ -1,7 +1,18 @@
 package org.partiql.lang.eval.physical.operators
 
 import org.partiql.lang.domains.PartiqlPhysical
+import org.partiql.lang.eval.ExprValue
+import org.partiql.lang.eval.ExprValueType
+import org.partiql.lang.eval.address
+import org.partiql.lang.eval.name
+import org.partiql.lang.eval.namedValue
+import org.partiql.lang.eval.physical.EvaluatorState
 import org.partiql.lang.eval.physical.SetVariableFunc
+import org.partiql.lang.eval.relation.RelationType
+import org.partiql.lang.eval.relation.relation
+import org.partiql.lang.eval.syntheticColumnName
+import org.partiql.lang.eval.unnamedValue
+import org.partiql.lang.planner.transforms.DEFAULT_IMPL_NAME
 
 /** Provides an implementation of the [PartiqlPhysical.Bexpr.Scan] operator.*/
 abstract class UnpivotRelationalOperatorFactory(name: String) : RelationalOperatorFactory {
@@ -24,3 +35,48 @@ abstract class UnpivotRelationalOperatorFactory(name: String) : RelationalOperat
         setByVar: SetVariableFunc?
     ): RelationExpression
 }
+
+internal class UnpivotOperator(name: String) : UnpivotRelationalOperatorFactory(name) {
+    override fun create(
+        impl: PartiqlPhysical.Impl,
+        expr: ValueExpression,
+        setAsVar: SetVariableFunc,
+        setAtVar: SetVariableFunc?,
+        setByVar: SetVariableFunc?
+    ): RelationExpression =
+        RelationExpression { state ->
+
+            fun ExprValue.unpivot(state: EvaluatorState): ExprValue = when (type) {
+                ExprValueType.STRUCT, ExprValueType.MISSING -> this
+
+                else -> state.valueFactory.newBag(
+                    listOf(
+                        this.namedValue(state.valueFactory.newString(syntheticColumnName(0)))
+                    )
+                )
+            }
+
+            val originalValue = expr(state)
+
+            val unpivot = originalValue.unpivot(state)
+
+            relation(RelationType.BAG) {
+                val iter = unpivot.iterator()
+                while (iter.hasNext()) {
+                    val item = iter.next()
+
+                    setAsVar(state, item.unnamedValue())
+
+                    if (setAtVar != null) {
+                        setAtVar(state, item.name ?: state.valueFactory.missingValue)
+                    }
+
+                    if (setByVar != null) {
+                        setByVar(state, item.address ?: state.valueFactory.missingValue)
+                    }
+                    yield()
+                }
+            }
+        }
+}
+
